// src/types/leverage-contracts/BalancerVault.ts
export declare namespace IVault {
  export type BatchSwapStepStruct = {
    poolId: BytesLike;
    assetInIndex: BigNumberish;
    assetOutIndex: BigNumberish;
    amount: BigNumberish;
    userData: BytesLike;
  };

  export type BatchSwapStepStructOutput = [
    poolId: string,
    assetInIndex: bigint,
    assetOutIndex: bigint,
    amount: bigint,
    userData: string
  ] & {
    poolId: string;
    assetInIndex: bigint;
    assetOutIndex: bigint;
    amount: bigint;
    userData: string;
  };

  export type FundManagementStruct = {
    sender: AddressLike;
    fromInternalBalance: boolean;
    recipient: AddressLike;
    toInternalBalance: boolean;
  };

  export type FundManagementStructOutput = [
    sender: string,
    fromInternalBalance: boolean,
    recipient: string,
    toInternalBalance: boolean
  ] & {
    sender: string;
    fromInternalBalance: boolean;
    recipient: string;
    toInternalBalance: boolean;
  };

  export type ExitPoolRequestStruct = {
    assets: AddressLike[];
    minAmountsOut: BigNumberish[];
    userData: BytesLike;
    toInternalBalance: boolean;
  };

  export type ExitPoolRequestStructOutput = [
    assets: string[],
    minAmountsOut: bigint[],
    userData: string,
    toInternalBalance: boolean
  ] & {
    assets: string[];
    minAmountsOut: bigint[];
    userData: string;
    toInternalBalance: boolean;
  };

  export type JoinPoolRequestStruct = {
    assets: AddressLike[];
    maxAmountsIn: BigNumberish[];
    userData: BytesLike;
    fromInternalBalance: boolean;
  };

  export type JoinPoolRequestStructOutput = [
    assets: string[],
    maxAmountsIn: bigint[],
    userData: string,
    fromInternalBalance: boolean
  ] & {
    assets: string[];
    maxAmountsIn: bigint[];
    userData: string;
    fromInternalBalance: boolean;
  };

  export type PoolBalanceOpStruct = {
    kind: BigNumberish;
    poolId: BytesLike;
    token: AddressLike;
    amount: BigNumberish;
  };

  export type PoolBalanceOpStructOutput = [
    kind: bigint,
    poolId: string,
    token: string,
    amount: bigint
  ] & { kind: bigint; poolId: string; token: string; amount: bigint };

  export type UserBalanceOpStruct = {
    kind: BigNumberish;
    asset: AddressLike;
    amount: BigNumberish;
    sender: AddressLike;
    recipient: AddressLike;
  };

  export type UserBalanceOpStructOutput = [
    kind: bigint,
    asset: string,
    amount: bigint,
    sender: string,
    recipient: string
  ] & {
    kind: bigint;
    asset: string;
    amount: bigint;
    sender: string;
    recipient: string;
  };

  export type SingleSwapStruct = {
    poolId: BytesLike;
    kind: BigNumberish;
    assetIn: AddressLike;
    assetOut: AddressLike;
    amount: BigNumberish;
    userData: BytesLike;
  };

  export type SingleSwapStructOutput = [
    poolId: string,
    kind: bigint,
    assetIn: string,
    assetOut: string,
    amount: bigint,
    userData: string
  ] & {
    poolId: string;
    kind: bigint;
    assetIn: string;
    assetOut: string;
    amount: bigint;
    userData: string;
  };
}